Update capter1.4 content
modify api: StaticSETofInts and ThreeSum

+++ API/ThreeSum.java
@@ -2,9 +2,6 @@ package com.leo.api;
 
 import java.util.Arrays;
 
-import com.leo.api.BinarySearch;
-import com.leo.api.Stopwatch;
-
 import edu.princeton.cs.algs4.StdRandom;
 
 public class ThreeSum {
@@ -74,7 +71,7 @@ public class ThreeSum {
 		for(int i=0; i< N; i++)
 			for(int j=i+1; j< N; j++)
 				for(int k=j+1; k< N; k++)
-					if(a[i]+a[j]+a[k]==0)
+					if((long)a[i]+a[j]+a[k]==0)
 						cnt++;
 		return cnt;
 	}
